chore(devql): split oversized sync modules
