+++ bitloops/src/host/devql/watch.rs
@@ -18,6 +18,14 @@ use crate::host::runtime_store::RepoSqliteRuntimeStore;
 
 #[path = "capture.rs"]
 mod capture;
+#[path = "watch_registration.rs"]
+mod registration;
+
+use registration::{
+    ExistingWatcherRegistrationHandle, TimedOutPendingRecovery, WatcherRegistrationReadyError,
+    handle_existing_watcher_registration, recover_timed_out_pending_registration,
+    wait_for_watcher_registration_ready,
+};
 
 const WATCHER_COMMAND_NAME: &str = "__devql-watcher";
 pub const DISABLE_WATCHER_AUTOSTART_ENV: &str = "BITLOOPS_DISABLE_WATCHER_AUTOSTART";
@@ -775,209 +783,6 @@ fn current_watcher_restart_token() -> Result<String> {
     let bytes = fs::read(&current_exe)
         .with_context(|| format!("reading watcher executable {}", current_exe.display()))?;
     Ok(hex::encode(Sha256::digest(bytes)))
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ExistingWatcherRegistrationDisposition {
-    Ready,
-    WaitForReady,
-    Replace { kill_running_process: bool },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ExistingWatcherRegistrationHandle {
-    Handled,
-    RetrySpawn,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum WatcherRegistrationReadyError {
-    ExitedBeforeReady { pid: u32 },
-    TimedOut { pid: u32 },
-}
-
-impl std::fmt::Display for WatcherRegistrationReadyError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ExitedBeforeReady { pid } => {
-                write!(
-                    f,
-                    "spawned DevQL watcher process {pid} exited before becoming ready"
-                )
-            }
-            Self::TimedOut { pid } => {
-                write!(
-                    f,
-                    "timed out waiting for DevQL watcher process {pid} to become ready"
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for WatcherRegistrationReadyError {}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TimedOutPendingRecovery {
-    PendingReleased,
-    ReadyRegistrationPresent,
-}
-
-fn classify_existing_watcher_registration(
-    entry: &crate::host::runtime_store::RepoWatcherRegistration,
-    expected_restart_token: &str,
-    watcher_running: bool,
-) -> ExistingWatcherRegistrationDisposition {
-    if watcher_running && entry.restart_token == expected_restart_token {
-        return match entry.state {
-            crate::host::runtime_store::RepoWatcherRegistrationState::Ready => {
-                ExistingWatcherRegistrationDisposition::Ready
-            }
-            crate::host::runtime_store::RepoWatcherRegistrationState::Pending => {
-                ExistingWatcherRegistrationDisposition::WaitForReady
-            }
-        };
-    }
-
-    ExistingWatcherRegistrationDisposition::Replace {
-        kill_running_process: watcher_running && entry.restart_token != expected_restart_token,
-    }
-}
-
-fn handle_existing_watcher_registration(
-    runtime_store: &RepoSqliteRuntimeStore,
-    entry: crate::host::runtime_store::RepoWatcherRegistration,
-    expected_restart_token: &str,
-    timeout: Duration,
-    poll_interval: Duration,
-) -> Result<ExistingWatcherRegistrationHandle> {
-    match classify_existing_watcher_registration(
-        &entry,
-        expected_restart_token,
-        process_is_running(entry.pid),
-    ) {
-        ExistingWatcherRegistrationDisposition::Ready => {
-            Ok(ExistingWatcherRegistrationHandle::Handled)
-        }
-        ExistingWatcherRegistrationDisposition::WaitForReady => {
-            match wait_for_watcher_registration_ready(
-                entry.pid,
-                expected_restart_token,
-                timeout,
-                poll_interval,
-                || runtime_store.load_watcher_registration(),
-                || Ok(process_is_running(entry.pid)),
-            ) {
-                Ok(()) => Ok(ExistingWatcherRegistrationHandle::Handled),
-                Err(wait_error)
-                    if matches!(
-                        wait_error.downcast_ref::<WatcherRegistrationReadyError>(),
-                        Some(WatcherRegistrationReadyError::ExitedBeforeReady { .. })
-                    ) =>
-                {
-                    runtime_store.delete_pending_watcher_registration_if_matches(
-                        entry.pid,
-                        &entry.restart_token,
-                    )?;
-                    Ok(ExistingWatcherRegistrationHandle::RetrySpawn)
-                }
-                Err(wait_error)
-                    if matches!(
-                        wait_error.downcast_ref::<WatcherRegistrationReadyError>(),
-                        Some(WatcherRegistrationReadyError::TimedOut { .. })
-                    ) =>
-                {
-                    match recover_timed_out_pending_registration(
-                        runtime_store,
-                        entry.pid,
-                        &entry.restart_token,
-                    )? {
-                        Some(TimedOutPendingRecovery::ReadyRegistrationPresent) => {
-                            Ok(ExistingWatcherRegistrationHandle::Handled)
-                        }
-                        Some(TimedOutPendingRecovery::PendingReleased) => {
-                            Ok(ExistingWatcherRegistrationHandle::RetrySpawn)
-                        }
-                        None => Err(wait_error),
-                    }
-                }
-                Err(wait_error) => Err(wait_error),
-            }
-        }
-        ExistingWatcherRegistrationDisposition::Replace {
-            kill_running_process,
-        } => {
-            if kill_running_process {
-                // Restart token mismatch means a different binary is now serving watcher work.
-                // Kill the stale watcher so the new process can re-run startup schema init.
-                kill_process(entry.pid);
-            }
-            runtime_store
-                .delete_watcher_registration_if_matches(entry.pid, &entry.restart_token)?;
-            Ok(ExistingWatcherRegistrationHandle::RetrySpawn)
-        }
-    }
-}
-
-fn recover_timed_out_pending_registration(
-    runtime_store: &RepoSqliteRuntimeStore,
-    pid: u32,
-    expected_restart_token: &str,
-) -> Result<Option<TimedOutPendingRecovery>> {
-    if runtime_store.delete_pending_watcher_registration_if_matches(pid, expected_restart_token)? {
-        return Ok(Some(TimedOutPendingRecovery::PendingReleased));
-    }
-
-    match runtime_store.load_watcher_registration()? {
-        Some(entry)
-            if entry.restart_token == expected_restart_token
-                && entry.state
-                    == crate::host::runtime_store::RepoWatcherRegistrationState::Ready
-                && process_is_running(entry.pid) =>
-        {
-            Ok(Some(TimedOutPendingRecovery::ReadyRegistrationPresent))
-        }
-        None => Ok(Some(TimedOutPendingRecovery::PendingReleased)),
-        Some(_) => Ok(None),
-    }
-}
-
-fn wait_for_watcher_registration_ready<FLoad, FWatcherRunning>(
-    expected_pid: u32,
-    expected_restart_token: &str,
-    timeout: Duration,
-    poll_interval: Duration,
-    mut load_registration: FLoad,
-    mut watcher_running: FWatcherRunning,
-) -> Result<()>
-where
-    FLoad: FnMut() -> Result<Option<crate::host::runtime_store::RepoWatcherRegistration>>,
-    FWatcherRunning: FnMut() -> Result<bool>,
-{
-    let deadline = Instant::now() + timeout;
-    loop {
-        if let Some(entry) = load_registration()?
-            && entry.pid == expected_pid
-            && entry.restart_token == expected_restart_token
-            && entry.state == crate::host::runtime_store::RepoWatcherRegistrationState::Ready
-        {
-            return Ok(());
-        }
-
-        if !watcher_running()? {
-            return Err(anyhow::Error::new(
-                WatcherRegistrationReadyError::ExitedBeforeReady { pid: expected_pid },
-            ));
-        }
-
-        if Instant::now() >= deadline {
-            return Err(anyhow::Error::new(
-                WatcherRegistrationReadyError::TimedOut { pid: expected_pid },
-            ));
-        }
-
-        std::thread::sleep(poll_interval);
-    }
 }
 
 async fn wait_for_shutdown_signal() {
